Issue #3891: re-organized api package inputs
