+++ src/test/resources/com/puppycrawl/tools/checkstyle/api/filetext/InputFileTextImportControl.xml
